main: make tool importable
by encapsulating everything related to
the actual calculation into a new function
called 'run', which returns a dict.

Also wrap the ArgParser with a function
called parse_args, to allow injecting
argument from a test.

And finally in addition set jobs per default
to the available cores of the machine.

Relates to #25

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>