pytest: switch to pytest-parallel
instead of using pytest-forked.
We can test that every run is really not using
global variables and we can running with all the
cores on a multicore system, heavily improving
our testing time.

Also switch back to pytest-cov 3.x, as 4.x removed
the support for getting coverage in a multiprocessing
application.

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>