License: relicense to zlib
relicense the repository to zLib license.
The amount of rights is almost the same as BSD-2-Clause.
But this license encourages to contribute back, instead of using
parts of it in a fork and redistributing it under a different name.

I think as a maintainer having one tool is better than having
a variation of it under different names.

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>