tokentree: add support for regex
to allow even more combinations instead of relying
only on plain string matching.
This will help if the tokenizer does only return complex
structures.

Also add include_start to config, to include the
starting token to the result.
Generalize the trim functionality and use replace
not strip for strings, as otherwise every character from
the string would be removed and this is not what
we want

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>